requirement.txt
asgiref==3.2.3
astroid==2.3.3
autopep8==1.5
bson==0.5.8
dataclasses==0.6
Django==3.0.3
django-cleanup==4.0.0
django-filter==2.2.0
djangorestframework==3.11.0
isort==4.3.21
lazy-object-proxy==1.4.3
Markdown==3.2.1
mccabe==0.6.1
Pillow==7.0.0
pycodestyle==2.5.0
pylint==2.4.4
pymongo==3.10.0
python-dateutil==2.8.1
pytz==2019.3
rope==0.16.0
six==1.14.0
sqlparse==0.2.4
typed-ast==1.4.1
wrapt==1.11.2
